Move :add_store_credit_payments to a service & add amount param
- Deprecate `add_store_credit_payments`
